Move visual_array_cell
